Move pathfinding functions to their own file
